Updated tests.

+++ tests/config.finder.test.js
@@ -2,47 +2,20 @@ var should = require('should'),
     finder = require('../platypi-cli/config/config.finder'),
     fs = require('fs');;
 
-describe('finder no config', function () {
+describe('finder', function () {
     it('should be a function', function () {
         finder.should.be.a.Function;
     });
 
-    it('should throw an exception', function () {
-        (function () {
-            finder();
-        }).should.throw();
-    });
-});
-
-describe('finder with config', function () {
-    var config = {
-        name: 'project',
-        type: 'mobile',
-        author: 'Donald Jones'
-    };
-
-    before(function (done) {
-        fs.writeFile('platypi.json', JSON.stringify(config), function (err) {
-            if (err) {
-                console.log(err);
-            }
-            done();
+    describe('no config', function () {
+        it('should throw an exception', function () {
+            (function () {
+                finder();
+            }).should.throw();
         });
     });
-    
-    it('should not throw an exception', function () {
-        (function () {
-            finder();
-        }).should.not.throw();
-    });
 
-    after(function () {
-        fs.unlink('platypi.json');
-    });
-});
-
-describe('finder with config in parent (recursive search)', function () {
-    if (!fs.existsSync('../platypi.json')) {
+    describe('with config', function () {
         var config = {
             name: 'project',
             type: 'mobile',
@@ -50,7 +23,7 @@ describe('finder with config in parent (recursive search)', function () {
         };
 
         before(function (done) {
-            fs.writeFile('../platypi.json', JSON.stringify(config), function (err) {
+            fs.writeFile('platypi.json', JSON.stringify(config), function (err) {
                 if (err) {
                     console.log(err);
                 }
@@ -65,14 +38,42 @@ describe('finder with config in parent (recursive search)', function () {
         });
 
         after(function () {
-            fs.unlink('../platypi.json');
+            fs.unlink('platypi.json');
         });
-    } else {
-        it('should not throw an exception', function () {
-            (function () {
-                finder();
-            }).should.not.throw();
-        });
-    }
+    });
 
+    describe('with config in parent (recursive search)', function () {
+        if (!fs.existsSync('../platypi.json')) {
+            var config = {
+                name: 'project',
+                type: 'mobile',
+                author: 'Donald Jones'
+            };
+
+            before(function (done) {
+                fs.writeFile('../platypi.json', JSON.stringify(config), function (err) {
+                    if (err) {
+                        console.log(err);
+                    }
+                    done();
+                });
+            });
+
+            it('should not throw an exception', function () {
+                (function () {
+                    finder();
+                }).should.not.throw();
+            });
+
+            after(function () {
+                fs.unlink('../platypi.json');
+            });
+        } else {
+            it('should not throw an exception', function () {
+                (function () {
+                    finder();
+                }).should.not.throw();
+            });
+        }
+    });
 });
